feat: Add support for default injection of envd in cubemastercli #0
